# swagger.yml
openapi: 3.0.0
info:
  title: School Management System
  description: This is a School Management System which provides automation of various academic and administrative operations.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://sch-management-server.herokuapp.com/api
    description: Production Server
paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Admin Signup
      requestBody:
        $ref: '#/components/requestBodies/AdminSignup'
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: Admin Created Successfully
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      requestBody:
        $ref: '#/components/requestBodies/Login'
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: User Logged In Successfully
  /fetchTeachers:
    get:
      tags:
        - Common
      summary: Fetch all the teachers
      responses:
        '200':
          description: Teachers Fetched Successfully
        '500':
          description: Internal Server Error
  /fetchStudents:
    get:
      tags:
        - Common
      summary: Fetch all the students
      responses:
        '200':
          description: Students Fetched Successfully
        '500':
          description: Internal Server Error
  /fetchClasses:
    get:
      tags:
        - Common
      summary: Fetch all the classes
      responses:
        '200':
          description: Classes Fetched Successfully
        '500':
          description: Internal Server Error
  /fetchBooks:
    get:
      tags:
        - Common
      summary: Fetch all the books
      responses:
        '200':
          description: Books Fetched Successfully
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
    AdminSignup:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SignupInput'
        description: Admin Signup
    Login:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginInput'
        description: Login
...